Modify NameContainsKeywordsPredicate to ContainsKeywordsPredicate to suit new find feature
